Added ExampleDocumentElement cache and 'Examples' section to element_detail.
We're now caching each XML element name that's used in
each ExampleDocument. This is updated automatically whenever
an ExampleDocument is added/changed.

The data is currently used on the element_detail page,
in a new 'Examples' section. This provides links to
all examples that use the given element.

Note that it's a bit imperfect because we're saving
the element name as a string as opposed to a ForeignKey
to XMLElement. This is because XMLElement objects might
change, and I didn't want to have to worry about cache
invalidation. As a consequence, this system isn't clever
enough to deal with elements such as <directions> that
can live in multiple places. This is an area for future
improvement (perhaps by storing the parent_element_name
in the ExampleDocumentElement model).