Mostly migrated MNX By Example to the new docs system.
The new docs now have a 'Comparing MNX and MusicXML' page,
which is essentially the same as the previous MNX By Example.
It lists all ExampleDocuments that have an associated MusicXML
example.

This necessitated creating two new models:

* DocumentFormat
* ExampleDocumentComparison

I've migrated all of the MusicXML examples from MNX By Example
into this new system, including the preambles.

The existing example_detail pages now link to /comparisons/musicxml/
if the given example has a MusicXML equivalent. (All ExampleDocuments
in the system currently do.)

Remaining tasks in order to bring this up to parity with MNX By Example:

* Add a way to specify the ordering of the examples on /comparisons/musicxml/.
At the moment ordering is undefined. Conveniently it happens to be the
same ordering as MNX By Example, but only because the examples were created
in the DB in the same order as MNX By Example (so this is not reliable
ordering long-term).

* Implement the 'Show relevant section' vs 'Show full document'. My
thinking here is that we add an XML attribute to the ExampleDocument
and ExampleDocumentComparison that specifies when each diff starts/ends.
Then we'd need to update utils.htmlutils.XMLPrettifier to output
some HTML markup around each diff.