Added proper ordering to 'Comparing MNX and MusicXML' page.
This is done with a new database column ExampleDocumentComparison.position.
There's no UI to set this value; it can be tweaked directly
in the DB via SQL commands using 'manage.py dbshell'.

Note that the page was already ordering the examples
correctly, but that was coincidental due to SQLite's
default/ambiguous ordering (because I created the
examples in the DB in a certain order).

This is the last step in migrating 'MNX By Example' to
the new docs system, so the old page can now be redirected
to this new page.