docgenerator: Added XMLElement base_element and is_abstract_element fields.
This makes it possible for this system to handle
various XSD types as used by the MusicXML schema.

The element_detail() view now properly accounts for
abstract elements when determining parent/child
elements. Essentially this makes abstract elements
'transparent' to people viewing the documentation;
they're treated as an implementation detail that
end users don't need to know or worry about.