docgenerator: Added SiteOptions model and changed site to use it.
This model is a singleton that contains general information
about the site. It currently has two fields:

* site_name -- e.g., 'MNX specification'
* xml_format_name -- e.g., 'MNX'

All templates now have access to the variable SITE_OPTIONS,
which is the SiteOptions object. As such, all templates now
use that instead of hard-coding 'MNX'. This will make it
possible for MusicXML to use this system, by changing the
SiteOptions.

The one remaining hard-coded 'MNX' is the 'Comparing MNX and MusicXML'
link on the homepage. That should be relatively easy to fix.