docgenerator: Added XMLElement.children_type field.
This isn't actually used anywhere in the doc generation,
but we'll use it for MusicXML.

In MNX, the order of children elements doesn't matter,
so this has been a core assumption of the docs system.
In MusicXML, order of children elements matters, so this
new field lets us specify that.