docgenerator: Added XMLSchema model.
This allows the docs system to handle multiple XML schemas,
which is required for the MusicXML docs.

Updated the XSD importer to work with XMLSchemas. As a result,
the command-line parameters to 'manage.py import_xsd' have
changed. The MusicXML section in the README is updated.

The XML schema data isn't yet surfaced anywhere in the website.