docgenerator: Fixed display error in example_detail markup.
The <metadiff> sections were being separated by <div>s,
which makes sense for the original use case in the
format_comparison page but doesn't make sense in the
context of example_detail.

The solution was to introduce a diffs_use_divs parameter,
which is set to False for example_detail pages. This uses
<span> instead of <div> for each <metadiff> -- meaning
the markup won't be visually broken up.