docgenerator: Improved XMLAugmenter to support xml:space=preserve.
This came up during the MusicXML docs migration. In the
'assess-and-player-elements' example, the <part-name>
element includes meaningful whitespace (a newline between
'Soprano' and 'Alto'), which we need to preserve.

The XMLAugmenter now looks for the xml:space='preserve'
attribute. If that's present, it preserves the whitespace
within the element's text content (except for leading and
trailing whitespace, which continue to be stripped; that
behavior is fine in this case).

This also necessitated changing XMLAugmenter to keep track
of current_characters, because it turns out the Python
SAX XML parser doesn't guarantee that every call to
characters() will contain the *entire* text content of an
element. The code was assuming that was the case. For an
XML element like this...

    <part-name>Soprano
Alto</part-name>

...the characters() method actually gets called three times,
with 'Soprano', '[newline]' and 'Alto'. We now handle that.