docgenerator: Fixed data_type_detail to account for abstract elements.
This had a bug when determining which elements use the
DataType as their content_data_type: it wasn't properly
accounting for abstract elements.

The problem was visible on the MusicXML 'accidental-value'
data type page, which showed no elements in the
'Where is this type used' section. It now properly displays
four elements there.