docgenerator: Improved ExampleDocumentElement cache to be more sophisticated.
Previously, ExampleDocumentElement simple cached the names
of all XML elements used in the example document (as strings).
This ended up being too naive of an approach, because some
XML element names are used by multiple elements in our system --
such as <part> in MusicXML (which has both partwise and timewise
versions).

This resulted in the problem where, for example, the
'<assess> and <player>' MusicXML example document was being
treated as a related document for the '<part> (timewise)'
element, whereas the document actually uses <part> (partwise).

The solution was to change the ExampleDocumentElement model to
have a foreign key to XMLElement, instead of just a string
element_name field. Then spec/utils/datautils.py had to be
tweaked, to be more sophisticated about which elements it
was seeing as it parsed the document (taking into account the
parent elements, via an element_stack); this is the same
approach we do in the XMLAugmenter, and we might want to
refactor in the future to remove duplication of the
get_element_obj() method (which is the same in both places).

Because the ExampleDocumentElement cache is only set when
an ExampleDocument is saved, any instances of this docs system
need to call save() on every ExampleDocument in the system
in order to refresh the cache.