docgenerator: Fixed ExampleDocuments to be scoped to schemas.
This fixes two bugs:

* On example_detail pages, the automatic linking of
elements to the appropriate pages wasn't working correctly
for the Opus schema in the MusicXML docs. This was because
the auto-linker (get_augmented_xml) wasn't filtering the
list of XMLElements to the example's schema; it was looking
at all XMLElements in the database.

* The ExampleDocumentElement cache generation also didn't
take schemas into account -- meaning the element_detail
'Examples' section was missing stuff for the MusicXML
Opus schema.