Moved main README content into docs, and regenerated docs.
It was confusing to have some documentation in the main GitHub
README and other documentation in the new docs tool. In this
commit, the content from the old README.md has been moved into
the new docs. The README now simply points to the docs.

This includes a few additions to the docs system (smufl-font-name
attribute, perform/visibility/display attributes, and a few small
other things -- see diff of data.json for details).