Renamed <directions> and <measure> elements to use distinct names in different contexts.
The changes are:

* <directions> (global) is now <directions-global>
* <directions> (part) is now <directions-part>
* <directions> (sequence) remains <directions>
* <measure> (global) is now <measure-global>
* <measure> (part) remains <measure>

All example documents have also been updated.

See #265 for discussion. The main rationale for this
change is as follows:

* Certain attributes (barline, index and number) only
make sense on <measure> (global), not on <measure> (part).
This implies the two elements should have different schema
definitions -- which is not allowed in W3C XML Schemas
if the elements have the same name and namespace.

* Having multiple elements with the same name in different
contexts would likely be confusing in the long term, both
for end users and for people developing the MNX spec.