docgenerator: Added support for JSON, instead of XML, as a schema's format.
A schema that's documented by this system can now be in JSON
format. Previously this system assumed every documented format
was XML.

The new models JSONObject and JSONObjectRelationship contain
the meat of the metadata about all of the objects and keys
within a JSON format. These models are roughly analogous to
the existing XMLElement and XMLRelationship models. I originally
tried to shoehorn JSON into the existing XMLElement model, but
that ended up being very confusing -- hence the totally separate
models.

There are two new Django views, json_object_list and json_object_detail.
These work similarly to the existing element_list and element_detail.

Example pages now properly support JSON formats, inserting the
appropriate deep links to doc pages for each element key/value.

The json_object_detail page also has an 'Examples' section, linking to
each example document in which that particular JSON object is used.

Note that this commit is purely the infrastructure for the docs system
and doesn't include the actual changes to MNX (to convert it from XML
into a JSON format). That will come in a subsequent commit.