Changed event's note value encoding to use an object.
Previously, an event's 'value' was a string using a microsyntax,
such as '/4' for quarter notes.

Now, the event uses a 'duration' key, whose value is a dictionary
with 'base' and 'dots' keys. The intent is to move away from
microsyntaxes to save every implementer from having to parse
them.

All example documents have been updated with the new encoding.