Added JSON Schema.
This updates the docgenerator system to be able to generate
a JSON Schema based on the data in the database.

A schema for MNX has been created, as docs/mnx-schema.json.
It's now linked-to from the left sidebar of each docs page.

Going forward, that schema will be automatically updated
each time we make any changes to the docs.