Changed accidental-display to use an 'enclosure' key.
Previously, accidental-display had two keys -- 'editorial'
and 'cautionary' -- to encode parentheses or square brackets.
These were a bit too ambiguous in retrospect, because they
were meant to be semantic but that left too much wiggleroom
for interpretation. We decided it makes more sense (and is much
more practical) if we just encode the pure graphical symbol
as opposed to its reasoning. In the future, we may expand this
to add a separate JSON key for 'reason' (editorial, cautionary,
etc.).

The accidental-display object now encodes this data in the
single key 'enclosure', which contains data about which symbol
is used to enclose the accidental.

This is deliberately a sub-object (as opposed to a simple string)
so that we can expand it in the future to encode the enclosure's
color, smufl font, position, reason, etc.