Changed docs system to serialize newline strings as lists of strings.
Previously, when serializing strings into our doc system's
JSON format, newlines were represented as the literal '\n'
character, resulting in very long lines that were hard to
read.

Now, we detect any strings with newlines in them, and we
convert these to lists of strings before serializing.
This means the JSON serialization is a bit easier to read.

This JSON file is purely an internal thing, so it's not
really intended to be human-readable, but this was a low-effort
way to make it a bit nicer.

Note that this is backwards compatible, so any JSON file
that was serialized in the old way will still be openable. It'll
be converted to the new format the next time it's serialized
(via 'manage.py freezedb').