Deleted spectools Django app and updated docs README.
The spectools package now lives in a separate GitHub project:
https://github.com/w3c-cg/mnxdocgenerator

This updates the README with new instructions on how to
install the spectools package, since it no longer lives
in this repo. That package hasn't yet been published to
PyPI, and we might decide not to do that (since it's such
a niche package and likely only used by two people). As
such, the installation instructions specifically use a
local pip installation.