Changed note.tie to note.ties.
This fixes a few edge cases of encoding ties, namely the fact
that a note might have multiple tie end notes (e.g., if it's
tied across multiple endings or a repeat/jump).

note.tie is now note.ties, an array of tie objects. The tie
object has been simplified to remove 'location' -- which
turned out to be unnecessary. To handle l.v. ties, we now
support an 'lv' key on the tie object.

Refs #396.