Changed event to make 'type':'event' optional.
Since events are by far the most common type of element within
a sequence, we dropped the requirement for them to define their
'type'. They still can, if software wants to be explicit, but
an element within a sequence is now assumed to be 'type':'event'
if the type isn't explicitly provided.

This also updates all example documents to remove 'type':'event'.