Added 'showOctave' key to clef objects.
This encodes whether a clef's glyph should display the octave
transposition.

The description for 'glyph' has changed to encourage people to
use 'showOctave' instead of 'glyph' if the only purpose of
encoding 'glyph' is to specify the octave-transposing clef glyph.

This also updates the multiple-layouts example document to use
showOctave=true and octave=-1.

Refs #424.