Allowed an optional 'id' in every object.
This uses the new mnxdocgenerator support for 'global attributes',
introduced in this commit:
https://github.com/w3c-cg/mnxdocgenerator/commit/e116d5e0b790bef0fcf5f4d586f1b5529f2ae17b

Every MNX object (technically every MNX object of type 'dictionary',
excluding 'user-defined dictionary') now supports an optional 'id'
key.

We had already supported 'id' on a small number of hand-chosen
objects, such as event. This commit removes those hand-coded 'id's
in favor of the new global 'id' attribute. (To be clear, this is
an implementation detail and hasn't changed the 'id' attribute on
those objects that already supported it.)